+++ docs/examples/03_backends_ros/files/04_attach_mesh_to_ee.py
@@ -23,4 +23,4 @@ with RosClient() as client:
     # sleep a bit before removing the tip
     time.sleep(1)
 
-    scene.remove_attached_collision_mesh('tip')
+    scene.reset()